move dailytable to sim folder
