+++ src/lib/simulation/dailyTable.ts
